+++ src/main/java/module/playerOverview/PlayerOverviewTable.java
@@ -142,7 +142,10 @@ public class PlayerOverviewTable extends JTable implements core.gui.Refreshable 
 		if (tableModel == null) {
 			tableModel = UserColumnController.instance().getPlayerOverviewModel();
 			tableModel.setValues(HOVerwaltung.instance().getModel().getAllSpieler());
-			tableSorter = new TableSorter(tableModel, tableModel.getDisplayedColumns().length - 1, getSortSpalte(), 0);
+			tableSorter = new TableSorter(tableModel,
+					tableModel.getPositionInArray(UserColumnFactory.ID),
+					getSortSpalte(),
+					tableModel.getPositionInArray(UserColumnFactory.NAME));
 
 			ToolTipHeader header = new ToolTipHeader(getColumnModel());
 			header.setToolTipStrings(tableModel.getTooltips());