+++ src/main/java/module/series/promotion/LeagueStatus.java
@@ -1,0 +1,10 @@
+package module.series.promotion;
+
+public enum LeagueStatus {
+
+    AVAILABLE,    // Data is available
+    IN_PROGRESS,  // Data being uploaded
+    BEING_PROCESSED, // Data uploaded, being processed.
+    NOT_AVAILABLE // Data is not available.
+
+}