+++ src/main/resources/release_notes.md
@@ -4,13 +4,23 @@ layout: page
 ---
 
 Changelist HO! 3.0
+
+## Squad
+  - Fix Sorting by Position #397
+
+## Lineup
+  - Make Name column resizable in players' table #382
+
+## League Details
+ - Promotion / Demotion status displayed on week 14 and 15 of season. #247
+
+
+Changelist HO! 2.1
 ====================
 
 
 ## Matches
   - Match Report: now similar to HT match report  #339
-  
+
 ## Player Analysis
   - fix age of player in player analysis table #422
-  
-